+++ libcoveocds/schema.py
@@ -62,7 +62,7 @@ class SchemaOCDS(SchemaJsonMixin):
         self.missing_package = False
         self.invalid_version_argument = False
         self.invalid_version_data = False
-        self.json_deref_error = None
+        self.json_deref_error = None  # str
 
         self._set_schema_version(self.config.config["schema_version"])